Split WeeklyWorklogSummaryUseCase test into focused modules
Split the large WeeklyWorklogSummaryUseCase.test.ts (1134 lines) into:
- WeeklyWorklogSummaryUseCase.testutils.ts (12 lines) - Shared test utilities
- WeeklyWorklogSummaryUseCase.basic.test.ts (397 lines) - Core functionality tests
- WeeklyWorklogSummaryUseCase.favorites.test.ts (141 lines) - Favorites integration tests
- WeeklyWorklogSummaryUseCase.sliding-window.test.ts (844 lines) - Sliding window logic tests

Total: 1394 lines (+260 lines due to test organization, but much more maintainable)

Benefits:
- Better test organization by feature area
- Easier to locate and modify specific test categories
- Shared test utilities reduce code duplication
- Maintains comprehensive test coverage (36 tests passed)
